requirements.txt
python-dotenv
pandas
numpy
certifi
pymongo[srv]
pymongo
scikit-learn
dill
pyaml
pickle
mlflow
##-e .